Move simulation data loading logic into sim folder
