+++ telnet/connection.go
@@ -9,8 +9,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/soupstore/coda-world/config"
-	"github.com/soupstore/coda-world/log"
+	"github.com/soupstore/coda-world/common/config"
+	"github.com/soupstore/coda-world/common/log"
 	"github.com/soupstore/coda-world/simulation"
 )
